Initial implementation of gravatar plugin
Adds avatar property to graphql User schema and a corresponding resolver to
create, if necessary, and return an avatar uuid.

Adds server route /avatar/:avatar to proxy avatar requests to Gravatar. It uses
redis cache to improve response times, and ETag http cache directive to enable
browser side caching.

Adds client side graphql container component and view component to render avatars.
Currently the Avatar slot is positioned so that the avatar ends up on top of the
comment rather than left of it. This could be fixed by external css stylesheet
floating the avatar slot and comment left (or something in that fashion).

Signed-off-by: Snorre Magnus Davøen <snorremd@gmail.com>